weather-data/requirements.txt
Flask==2.0.3
requests==2.26.0
boto3==1.20.24
flask_cors==3.0.10
Werkzeug==2.0.3
gunicorn==20.1.0
mysql-connector-python==8.0.28 
